fix: fix comment and post features

+++ app/Http/Controllers/api/v1/CommentController.php
@@ -3,39 +3,71 @@
 namespace App\Http\Controllers\api\V1;
 
 use App\Models\Comment;
+use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreCommentRequest;
-use App\Http\Requests\UpdateCommentRequest;
 
 class CommentController extends Controller
 {
     // Store a newly created resource in storage.
-    public function store(StoreCommentRequest $request)
+    public function store(Request $request)
     {
-        $validated = $request->validate(['comment_description' => 'required']);
+        $validated = $request->validate([
+            'user_id' => 'required',
+            'post_id' => 'required',
+            'comment_description' => 'required'
+        ]);
 
-        Comment::create(['comment_description' => $validated['comment_desciption']]);
+        Comment::create([
+            'user_id' => $validated['user_id'],
+            'post_id' => $validated['post_id'],
+            'comment_description' => $validated['comment_description']
+        ]);
 
-        return response()->json(['message' => "Comment created successfully!"], 201);
+        return response()->json([
+            'message' => "Comment created successfully!"
+        ], 201);
     }
 
     // Update the specified resource in storage.
-    public function update(UpdateCommentRequest $request, Comment $comment)
+    public function update(Request $request, string $id)
     {
-        $validated = $request->validate(['comment_description' => 'required']);
+        $comment = Comment::find($id) ?? null;
 
-        $comment->update(['comment_description' => $validated['comment_description']]);
+        if($comment) {
+            $validated = $request->validate(['comment_description' => 'required']);
 
-        return response()->json(['message' => "Comment updated successfully!"]);
+            $comment->update([
+                'comment_description' => $validated['comment_description']
+            ]);
+
+            return response()->json([
+                'message' => "Comment updated successfully!"
+            ]);
+        } else {
+            return response()->json([
+                'error' => "Comment not found",
+            ], 404); 
+        }
+        
     }
 
     // Remove the specified resource from storage.
-    public function destroy(Comment $comment)
+    public function destroy(string $id)
     {
-        $comment->delete();
+        $comment = Comment::find($id) ?? null;
 
-        return response()->json([
-            'message' => 'Comment deleted successfully',
-        ]);
+        if($comment) {
+            
+            $comment->delete();
+
+            return response()->json([
+                'message' => 'Comment deleted successfully',
+            ]);
+
+        } else {
+            return response()->json([
+                'error' => "Comment not found",
+            ], 404); 
+        }
     }
 }
